added error authentication layer for delete

+++ routes/story.js
@@ -52,8 +52,20 @@ router.put('/:id', ensureAuthentication, urlencodedParser, async(req, res) => {
 router.delete('/:id', ensureAuthentication, async(req, res) => {
     try {
         const {id} = req.params;
-        await Story.findByIdAndDelete(id);
-        res.redirect('/dashboard');
+        const story = Story.findById(id);
+
+        if(!story){
+            throw new ApplicationError("Story trying to delete doesn't exist", 404);
+        }
+        
+        if(req.user.id != story.user){
+            return res.redirect('/stories');
+        } else {
+            await Story.findByIdAndDelete(id);
+            res.redirect('/dashboard');
+        }
+
+
     } catch (error) {
         console.log(error);
     }
